+++ malachite-nz/src/integer/logic/bit_convertible.rs
@@ -106,7 +106,7 @@ pub fn bits_vec_to_twos_complement_bits_negative(bits: &mut Vec<bool>) {
 fn limbs_asc_from_negative_twos_complement_limbs_asc(mut limbs: Vec<Limb>) -> Vec<Limb> {
     {
         let most_significant_limb = limbs.last_mut().unwrap();
-        let leading_zeros = most_significant_limb.leading_zeros();
+        let leading_zeros = u64::from(most_significant_limb.leading_zeros());
         if leading_zeros != 0 {
             *most_significant_limb |= !((1 << (Limb::WIDTH - leading_zeros)) - 1);
         }
