+++ malachite-test/src/tune/div_mod_divide_and_conquer_to_barrett.rs
@@ -14,7 +14,7 @@ pub(crate) fn tune() -> Vec<String> {
     let result = compare_two(
         &mut (|(mut qs, mut ns, mut ds): (Vec<Limb>, Vec<Limb>, Vec<Limb>)| {
             let q_len = ns.len() - ds.len() + 1;
-            ds[q_len - 1].set_bit(u64::from(Limb::WIDTH) - 1);
+            ds[q_len - 1].set_bit(Limb::WIDTH - 1);
             let inverse = limbs_two_limb_inverse_helper(ds[q_len - 1], ds[q_len - 2]);
             no_out!(_limbs_div_mod_divide_and_conquer(
                 &mut qs,
@@ -28,7 +28,7 @@ pub(crate) fn tune() -> Vec<String> {
             let mut rs = vec![0; d_len];
             let q_len = ns.len() - d_len + 1;
             let q_len_2 = q_len << 1;
-            ds[q_len - 1].set_bit(u64::from(Limb::WIDTH) - 1);
+            ds[q_len - 1].set_bit(Limb::WIDTH - 1);
             limbs_two_limb_inverse_helper(ds[q_len - 1], ds[q_len - 2]);
             let mut scratch = vec![0; _limbs_div_mod_barrett_scratch_len(q_len_2, q_len)];
             _limbs_div_mod_barrett(&mut qs, &mut rs, &ns[..q_len_2], &ds[..q_len], &mut scratch);
